requirements.txt
jiwer==2.5.1
lightning-bolts==0.5.0
pandas==1.5.0
pytorch-lightning==1.7.7
pytorch-lightning-bolts==0.3.2.post1
PyYAML==5.4.1
torch==1.12.1
torchaudio==0.12.1
torchmetrics==0.9.3
torchvision==0.13.1
tqdm==4.64.1